requirements.txt
Flask==0.12.1
transitions==0.5.0
python-telegram-bot==5.3.0
pygraphviz==1.3.1
requests==2.11.1
beautifulsoup4==4.5.1